Add whenCalledOrAfter() to ensure callbacks are called
whenCalledOrAfter() wraps a callback function and calls it after
a specified number of milliseconds if it's not already been called.

The current use case is to wrap GA hitCallbacks for important things
like page redirects after an event has been acknowledged. e.g. we have
to wait for the event to send before redirecting otherwise it wouldn't
always hit GA, but we can't wait forever for GA to acknowledge it, so we
need to give up after a short time.