Enable the linkid Google Analytics extension
This is needed to track which link on a page sent a user to a
destination page in the case that there are multiple links to the same
page on a page. This happens in Avocet with our many links to the
upload page - we're interested in knowing how much each are used.